backend-flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
web3==6.14.0
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
